+++ test/unit/fuzz-tools.test.js
@@ -313,6 +313,7 @@ import * as kikuchiV333 from '../../lib/kikuchi-v333.js';
 import * as kudoV334 from '../../lib/kudo-v334.js';
 import * as niceV335 from '../../lib/nice-v335.js';
 import * as jnetV336 from '../../lib/jnet-v336.js';
+import * as outerbridgeV337 from '../../lib/outerbridge-v337.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -571,6 +572,7 @@ const MODULES = {
   'kudo-v334.js': kudoV334,
   'nice-v335.js': niceV335,
   'jnet-v336.js': jnetV336,
+  'outerbridge-v337.js': outerbridgeV337,
   'tb-testing.js': tbTesting,
 };